+++ src/OpenTelemetry/CHANGELOG.md
@@ -9,6 +9,13 @@
 * Make `MetricPoint` of `MetricPointAccessor` readonly.
   ([#2736](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2736))
 
+* Fail-fast when using AddView with guaranteed conflict.
+  ([2751](https://github.com/open-telemetry/opentelemetry-dotnet/issues/2751))
+
+* Swallow `ObjectDisposedException` from the `BatchExportProcessor` worker
+  thread.
+  ([2844](https://github.com/open-telemetry/opentelemetry-dotnet/issues/2844))
+
 ## 1.2.0-rc1
 
 Released 2021-Nov-29